removed cast to float for cosAngularDistance Original: 97a6420ab86ff72b8c5308b980cca6a232ed6be9

+++ lalapps/src/pulsar/FDS_isolated/polka.c
@@ -234,7 +234,7 @@ int main(int argc,char *argv[])
 	  if (cosAngularDistance  >  1.0) cosAngularDistance =  1.0;
 	  if (cosAngularDistance  < -1.0) cosAngularDistance = -1.0;
 
-	  AngularDistance=acos((float) cosAngularDistance);
+	  AngularDistance=acos(cosAngularDistance);
 
 	  difff=fabs(f1 - CList2.f[indices2f[f]]);
